# Conflicts:
#	kura/distrib/config/kura.build.properties
#	kura/distrib/pom.xml
#	kura/distrib/src/main/ant/build_equinox_distrib.xml
#	kura/emulator/org.eclipse.kura.emulator.net/OSGI-INF/network.xml
#	kura/org.eclipse.kura.core/src/main/java/org/eclipse/kura/core/system/SystemAdminServiceImpl.java
#	kura/org.eclipse.kura.emulator/OSGI-INF/network.xml
#	kura/org.eclipse.kura.windows.command/OSGI-INF/command.xml
#	target-platform/p2-repo-common/pom.xml
#	target-platform/pom.xml

+++ kura/org.eclipse.kura.core.deployment/src/main/java/org/eclipse/kura/core/deployment/install/InstallImpl.java
@@ -28,6 +28,10 @@ import org.slf4j.LoggerFactory;
 
 public class InstallImpl {
 	private static final Logger s_logger = LoggerFactory.getLogger(InstallImpl.class);
+	
+	private static final int   PROGRESS_COMPLETE = 100;
+	private static final String MESSAGE_CONFIGURATION_FILE_NOT_SPECIFIED = "Configuration file not specified";
+
 	public static final String RESOURCE_INSTALL = "install";
 
 	public static final String PERSISTANCE_SUFFIX = "_persistance";
@@ -49,13 +53,21 @@ public class InstallImpl {
 		this.callback = callback;
 
 		deployedPackages = new Properties();
-		m_installPersistanceDir= kuraDataDir + File.separator + PERSISTANCE_FOLDER_NAME;
+		StringBuilder pathSB= new StringBuilder();
+		pathSB.append(kuraDataDir);
+		pathSB.append(File.separator);
+		pathSB.append(PERSISTANCE_FOLDER_NAME);
+		m_installPersistanceDir= pathSB.toString();
 		File installPersistanceDir = new File(m_installPersistanceDir);
 		if (!installPersistanceDir.exists()) {
 			installPersistanceDir.mkdir();
 		}
 
-		m_installVerifDir= m_installPersistanceDir + File.separator + PERSISTANCE_VERIFICATION_FOLDER_NAME;
+		pathSB= new StringBuilder();
+		pathSB.append(m_installPersistanceDir);
+		pathSB.append(File.separator);
+		pathSB.append(PERSISTANCE_VERIFICATION_FOLDER_NAME);
+		m_installVerifDir= pathSB.toString();
 		File installVerificationDir = new File(m_installVerifDir);
 		if (!installVerificationDir.exists()) {
 			installVerificationDir.mkdir();
@@ -89,7 +101,7 @@ public class InstallImpl {
 	public void installDp(DeploymentPackageInstallOptions options, File dpFile) throws KuraException{
 		SafeProcess proc = null;
 		try {
-			installDeploymentPackageInternal(dpFile, options);
+			installDeploymentPackageInternal(dpFile);
 			installCompleteAsync(options, dpFile.getName());
 			s_logger.info("Install completed!");
 
@@ -101,8 +113,9 @@ public class InstallImpl {
 			s_logger.info("Install failed!");
 			installFailedAsync(options, dpFile.getName(), e);
 		} finally {
-			if (proc != null) ProcessUtil.destroy(proc);
-			dpFile = null;
+			if (proc != null) {
+				ProcessUtil.destroy(proc);
+			}
 		}
 	}
 
@@ -117,7 +130,9 @@ public class InstallImpl {
 		} catch (IOException e) {
 			throw new KuraException(KuraErrorCode.INTERNAL_ERROR);
 		} finally {
-			if (proc != null) ProcessUtil.destroy(proc);
+			if (proc != null) {
+				ProcessUtil.destroy(proc);
+			}
 		}
 
 		SafeProcess proc2 = null;
@@ -126,7 +141,9 @@ public class InstallImpl {
 		} catch (IOException e) {
 			throw new KuraException(KuraErrorCode.INTERNAL_ERROR);
 		} finally {
-			if (proc2 != null) ProcessUtil.destroy(proc2);
+			if (proc2 != null) {
+				ProcessUtil.destroy(proc2);
+			}
 		}
 	}
 
@@ -143,22 +160,18 @@ public class InstallImpl {
 	}
 
 	public void installCompleteAsync(DeploymentPackageOptions options, String dpName) throws KuraException{
-		KuraInstallPayload notify = null;
-
-		notify = new KuraInstallPayload(options.getClientId());
+		KuraInstallPayload notify = new KuraInstallPayload(options.getClientId());
 		notify.setTimestamp(new Date());
 		notify.setInstallStatus(INSTALL_STATUS.COMPLETED.getStatusString());
 		notify.setJobId(options.getJobId());
 		notify.setDpName(dpName); //Probably split dpName and dpVersion?
-		notify.setInstallProgress(100);
+		notify.setInstallProgress(PROGRESS_COMPLETE);
 
 		callback.publishMessage(options, notify, RESOURCE_INSTALL);
 	}
 
 	public void installFailedAsync(DeploymentPackageInstallOptions options, String dpName, Exception e) throws KuraException{
-		KuraInstallPayload notify = null;
-
-		notify = new KuraInstallPayload(options.getClientId());
+		KuraInstallPayload notify = new KuraInstallPayload(options.getClientId());
 		notify.setTimestamp(new Date());
 		notify.setInstallStatus(INSTALL_STATUS.FAILED.getStatusString());
 		notify.setJobId(options.getJobId());
@@ -183,7 +196,9 @@ public class InstallImpl {
 					} catch (IOException e) {
 
 					} finally {
-						if (proc != null) ProcessUtil.destroy(proc);
+						if (proc != null) {
+							ProcessUtil.destroy(proc);
+						}
 					}
 
 					SafeProcess proc2 = null;
@@ -191,15 +206,17 @@ public class InstallImpl {
 						proc2 = ProcessUtil.exec(fileEntry.getCanonicalPath());
 						int exitValue = proc2.exitValue();
 						if(exitValue == 0){
-							sendSysUpdateSuccess(fileEntry);
+							sendSysUpdateSuccess();
 						} else {
-							sendSysUpdateFailure(fileEntry);
+							sendSysUpdateFailure();
 						}
 					} catch (Exception e) {
 
 					} finally {
 						fileEntry.delete();
-						if (proc2 != null) ProcessUtil.destroy(proc2);
+						if (proc2 != null) {
+							ProcessUtil.destroy(proc2);
+						}
 					}
 
 				}
@@ -207,7 +224,7 @@ public class InstallImpl {
 		}
 	}
 
-	private DeploymentPackage installDeploymentPackageInternal(File fileReference, DeploymentPackageOptions options) 
+	private DeploymentPackage installDeploymentPackageInternal(File fileReference) 
 			throws DeploymentException, IOException {
 
 		InputStream dpInputStream = null;
@@ -217,9 +234,12 @@ public class InstallImpl {
 
 		try {
 			String dpBasename = fileReference.getName();
-			String dpPersistentFilePath = packagesPath + File.separator + dpBasename;
+			StringBuilder pathSB= new StringBuilder();
+			pathSB.append(packagesPath);
+			pathSB.append(File.separator);
+			pathSB.append(dpBasename);
+			String dpPersistentFilePath = pathSB.toString();
 			dpPersistentFile = new File(dpPersistentFilePath);
-			//downloadedFile = getDpDownloadFile(options);
 
 
 			dpInputStream = new FileInputStream(downloadedFile);
@@ -248,7 +268,10 @@ public class InstallImpl {
 			}
 			// The file from which we have installed the deployment package will be deleted
 			// unless it's a persistent deployment package file.
-			if (downloadedFile != null && !downloadedFile.getCanonicalPath().equals(dpPersistentFile.getCanonicalPath())) {
+			if (    downloadedFile != null && 
+					dpPersistentFile != null && 
+					!downloadedFile.getCanonicalPath().equals(dpPersistentFile.getCanonicalPath())
+					) {
 				downloadedFile.delete();
 			}			
 		}
@@ -266,13 +289,18 @@ public class InstallImpl {
 		m_installPersistance.setProperty(PERSISTANCE_FILE_NAME, fileName);
 
 		if (m_installPersistanceDir == null) {
-			s_logger.warn("Configuration file not specified");
+			s_logger.warn(MESSAGE_CONFIGURATION_FILE_NOT_SPECIFIED);
 			return;
 		}
 
 		FileOutputStream fos= null;
 		try {
-			String persistanceFile= m_installPersistanceDir + File.separator + fileName + PERSISTANCE_SUFFIX;
+			StringBuilder pathSB= new StringBuilder();
+			pathSB.append(m_installPersistanceDir);
+			pathSB.append(File.separator);
+			pathSB.append(fileName);
+			pathSB.append(PERSISTANCE_SUFFIX);
+			String persistanceFile= pathSB.toString();
 			fos = new FileOutputStream(persistanceFile);
 			m_installPersistance.store(fos, null);
 			fos.flush();
@@ -284,7 +312,7 @@ public class InstallImpl {
 				try {
 					fos.close();
 				} catch (IOException e) {
-					e.printStackTrace();
+					s_logger.error("Exception while closing the stream!", e);
 				}
 			}
 		}
@@ -294,18 +322,26 @@ public class InstallImpl {
 		deployedPackages.setProperty(packageName, packageUrl);
 
 		if (dpaConfPath == null) {
-			s_logger.warn("Configuration file not specified");
+			s_logger.warn(MESSAGE_CONFIGURATION_FILE_NOT_SPECIFIED);
 			return;
 		}
 
+		FileOutputStream fos = null;
 		try {
-			FileOutputStream fos = new FileOutputStream(dpaConfPath);
+			fos = new FileOutputStream(dpaConfPath);
 			deployedPackages.store(fos, null);
 			fos.flush();
 			fos.getFD().sync();
-			fos.close();
 		} catch (IOException e) {
 			s_logger.error("Error writing package configuration file", e);
+		} finally {
+			try{
+				if (fos != null){
+					fos.close();
+				}
+			} catch (IOException e){
+				s_logger.error("Exception while closing opened resources!", e);
+			}
 		}
 	}
 
@@ -313,78 +349,91 @@ public class InstallImpl {
 		deployedPackages.remove(packageName);
 
 		if (dpaConfPath == null) {
-			s_logger.warn("Configuration file not specified");
+			s_logger.warn(MESSAGE_CONFIGURATION_FILE_NOT_SPECIFIED);
 			return;
 		}
 
+		FileOutputStream fos= null;
 		try {
-			FileOutputStream fos = new FileOutputStream(dpaConfPath);
+			fos = new FileOutputStream(dpaConfPath);
 			deployedPackages.store(fos, null);
 			fos.flush();
 			fos.getFD().sync();
 			fos.close();
 		} catch (IOException e) {
 			s_logger.error("Error writing package configuration file", e);
+		} finally {
+			try{
+				if (fos != null){
+					fos.close();
+				}
+			} catch (IOException e){
+				s_logger.error("Exception while closing opened resources!", e);
+			}
 		}
 	}
 
-	private void sendSysUpdateSuccess(File verificationFile) throws KuraException {
+	private void sendSysUpdateSuccess() throws KuraException {
 		s_logger.info("Ready to send success after install");
 		File installDir= new File(m_installPersistanceDir);
-		for (File fileEntry : installDir.listFiles()) {
+		if(installDir != null){
+			for (File fileEntry : installDir.listFiles()) {
 
-			if (fileEntry.isFile() && fileEntry.getName().endsWith(PERSISTANCE_SUFFIX)) { //&& fileEntry.getName().contains(verificationFile.getName()
-				Properties downloadProperties= loadInstallPersistance(fileEntry);
-				String deployUrl= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_DOWNLOAD_URI);
-				String dpName= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_NAME);
-				String dpVersion= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_VERSION);
-				String clientId= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_CLIENT_ID);
-				Long jobId= Long.valueOf(downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_JOB_ID));
-				String fileSystemFileName= downloadProperties.getProperty(PERSISTANCE_FILE_NAME);
-				String requestClientId = downloadProperties.getProperty(CloudDeploymentHandlerV2.METRIC_REQUESTER_CLIENT_ID);
+				if (fileEntry.isFile() && fileEntry.getName().endsWith(PERSISTANCE_SUFFIX)) { //&& fileEntry.getName().contains(verificationFile.getName()
+					Properties downloadProperties= loadInstallPersistance(fileEntry);
+					String deployUrl= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_DOWNLOAD_URI);
+					String dpName= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_NAME);
+					String dpVersion= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_VERSION);
+					String clientId= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_CLIENT_ID);
+					Long jobId= Long.valueOf(downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_JOB_ID));
+					String fileSystemFileName= downloadProperties.getProperty(PERSISTANCE_FILE_NAME);
+					String requestClientId = downloadProperties.getProperty(CloudDeploymentHandlerV2.METRIC_REQUESTER_CLIENT_ID);
 
-				DeploymentPackageDownloadOptions options = new DeploymentPackageDownloadOptions(deployUrl, dpName, dpVersion);
-				options.setClientId(clientId);
-				options.setJobId(jobId);
-				options.setRequestClientId(requestClientId);
+					DeploymentPackageDownloadOptions deployOptions = new DeploymentPackageDownloadOptions(deployUrl, dpName, dpVersion);
+					deployOptions.setClientId(clientId);
+					deployOptions.setJobId(jobId);
+					deployOptions.setRequestClientId(requestClientId);
 
-				try {
-					installCompleteAsync(options, fileSystemFileName);
-					s_logger.info("Sent install complete");
-					fileEntry.delete();
-					break;
-				} catch (KuraException e) {
-					throw new KuraException(KuraErrorCode.INTERNAL_ERROR);
+					try {
+						installCompleteAsync(deployOptions, fileSystemFileName);
+						s_logger.info("Sent install complete");
+						fileEntry.delete();
+						break;
+					} catch (KuraException e) {
+						throw new KuraException(KuraErrorCode.INTERNAL_ERROR);
+					}
 				}
 			}
 		}
 	}
 
-	private void sendSysUpdateFailure(File verificationFile) throws KuraException {
+	private void sendSysUpdateFailure() throws KuraException {
 		File installDir= new File(m_installPersistanceDir);
-		for (final File fileEntry : installDir.listFiles()) {
-			if (fileEntry.isFile() && fileEntry.getName().endsWith(PERSISTANCE_SUFFIX)) { //&& fileEntry.getName().contains(verificationFile.getName())
-				Properties downloadProperties= loadInstallPersistance(fileEntry);
-				String deployUrl= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_DOWNLOAD_URI);
-				String dpName= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_NAME);
-				String dpVersion= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_VERSION);
-				String clientId= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_CLIENT_ID);
-				Long jobId= Long.valueOf(downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_JOB_ID));
-				String fileSystemFileName= downloadProperties.getProperty(PERSISTANCE_FILE_NAME);
-				String requestClientId = downloadProperties.getProperty(CloudDeploymentHandlerV2.METRIC_REQUESTER_CLIENT_ID);
+		if(installDir != null){
+			for (final File fileEntry : installDir.listFiles()) {
+				if (fileEntry.isFile() && fileEntry.getName().endsWith(PERSISTANCE_SUFFIX)) { //&& fileEntry.getName().contains(verificationFile.getName())
+					Properties downloadProperties= loadInstallPersistance(fileEntry);
+					String deployUrl= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_DOWNLOAD_URI);
+					String dpName= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_NAME);
+					String dpVersion= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_VERSION);
+					String clientId= downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_DP_CLIENT_ID);
+					Long jobId= Long.valueOf(downloadProperties.getProperty(DeploymentPackageDownloadOptions.METRIC_JOB_ID));
+					String fileSystemFileName= downloadProperties.getProperty(PERSISTANCE_FILE_NAME);
+					String requestClientId = downloadProperties.getProperty(CloudDeploymentHandlerV2.METRIC_REQUESTER_CLIENT_ID);
 
-				DeploymentPackageDownloadOptions options = new DeploymentPackageDownloadOptions(deployUrl, dpName, dpVersion);
-				options.setClientId(clientId);
-				options.setJobId(jobId);
-				options.setRequestClientId(requestClientId);
+					DeploymentPackageDownloadOptions deployOptions = new DeploymentPackageDownloadOptions(deployUrl, dpName, dpVersion);
+					deployOptions.setClientId(clientId);
+					deployOptions.setJobId(jobId);
+					deployOptions.setRequestClientId(requestClientId);
 
-				try {
-					installFailedAsync(options, fileSystemFileName, new KuraException(KuraErrorCode.INTERNAL_ERROR));
-					s_logger.info("Sent install failed");
-					fileEntry.delete();
-					break;
-				} catch (KuraException e) {
-					throw new KuraException(KuraErrorCode.INTERNAL_ERROR);
+					try {
+						installFailedAsync(deployOptions, fileSystemFileName, new KuraException(KuraErrorCode.INTERNAL_ERROR));
+						s_logger.info("Sent install failed");
+						fileEntry.delete();
+						break;
+					} catch (KuraException e) {
+						throw new KuraException(KuraErrorCode.INTERNAL_ERROR);
+					}
 				}
 			}
 		}
@@ -392,10 +441,20 @@ public class InstallImpl {
 
 	private Properties loadInstallPersistance(File installedDpPersistance){
 		Properties downloadProperies= new Properties();
+		FileReader fr = null;
 		try {
-			downloadProperies.load(new FileReader(installedDpPersistance));
+			fr= new FileReader(installedDpPersistance);
+			downloadProperies.load(fr);
 		} catch (IOException e) {
 			s_logger.error("Exception loading install configuration file", e);
+		} finally {
+			try{
+				if (fr != null) {
+					fr.close();
+				}
+			} catch (IOException e){
+				s_logger.error("Exception while closing opened resources!", e);
+			}
 		}
 		return downloadProperies;
 	}
