+++ kura/org.eclipse.kura.linux.net/src/main/java/org/eclipse/kura/linux/net/ppp/PppLinux.java
@@ -57,8 +57,9 @@ public class PppLinux {
 		int pid = getPid(iface, port);
 		if(pid >= 0) {
     		s_logger.info("killing {}  pid={}", iface, pid);
-    		LinuxProcessUtil.kill(pid);
-    		
+    		if (!LinuxProcessUtil.stop(pid)) {
+    			LinuxProcessUtil.kill(pid);
+    		}
     		if (port.startsWith("/dev/")) {
     			port = port.substring("/dev/".length());
     		}
